Add temporary password column and move auth env vars to main config.py
