+++ core_backend/app/config.py
@@ -5,6 +5,16 @@ LOG_LEVEL = os.environ.get("LOG_LEVEL", "INFO")
 # Root domain
 DOMAIN = os.environ.get("DOMAIN", "localhost")
 
+# user 1
+USER1_USERNAME = os.environ.get("USER1_USERNAME", "user1")
+USER1_PASSWORD = os.environ.get("USER1_PASSWORD", "fullaccess")
+USER1_RETRIEVAL_KEY = os.environ.get("USER1_RETRIEVAL_KEY", "user1-key")
+
+# user 2
+USER2_USERNAME = os.environ.get("USER2_USERNAME", "user2")
+USER2_PASSWORD = os.environ.get("USER2_PASSWORD", "fullaccess")
+USER2_RETRIEVAL_KEY = os.environ.get("USER2_RETRIEVAL_KEY", "user2-key")
+
 # Postgres variables
 POSTGRES_USER = os.environ.get("POSTGRES_USER", "postgres")
 POSTGRES_PASSWORD = os.environ.get("POSTGRES_PASSWORD", "postgres")
